use FilterLabel component
